backend/requirements.txt
fastapi
uvicorn[standard]
scikit-learn
pandas
numpy
joblib
shap
python-multipart
xgboost
imbalanced-learn
tensorflow
Pillow
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
psycopg2-binary
pytesseract
opencv-python

